// apps/bridge/codegen/model/forcedWithdrawalDto.ts
/**
 * Generated by orval v6.22.1 🍺
 * Do not edit manually.
 * API
 * API docs
 * OpenAPI spec version: 1.0
 */
import type { PortalDepositDto } from './portalDepositDto';
import type { OptimismTransactionType } from './optimismTransactionType';
import type { BridgeWithdrawalDto } from './bridgeWithdrawalDto';

export interface ForcedWithdrawalDto {
  deposit: PortalDepositDto;
  id: string;
  type: OptimismTransactionType;
  withdrawal?: BridgeWithdrawalDto;
}